fix(amf): addressed Github actions

+++ lte/gateway/c/core/oai/tasks/amf/amf_app_timer_management.h
@@ -30,7 +30,7 @@ namespace magma5g {
 typedef uint32_t timer_arg_t;
 // typedef std::pair<uint8_t, uint8_t> ue_pdu_id;
 typedef struct ue_pdu_id {
-  uint8_t ue_id;
+  uint32_t ue_id;
   uint8_t pdu_id;
 } ue_pdu_id_t;
